Add setup for removing roles on mute

+++ cogs/bot_settings.py
@@ -193,9 +193,10 @@ class BotSettings(utils.Cog):
                 embed.description = '\n'.join([
                     "1\N{COMBINING ENCLOSING KEYCAP} Set verified role (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("verified_role_id")) or MentionableNone("none")),
                     "2\N{COMBINING ENCLOSING KEYCAP} Set mute role (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("muted_role_id")) or MentionableNone("none")),
-                    "3\N{COMBINING ENCLOSING KEYCAP} Set moderator role (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("guild_moderator_role_id")) or MentionableNone("none")),
-                    "4\N{COMBINING ENCLOSING KEYCAP} Set custom role master (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("custom_role_id")) or MentionableNone("none")),
-                    "5\N{COMBINING ENCLOSING KEYCAP} Set custom role position (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("custom_role_position_id")) or MentionableNone("none")),
+                    "3\N{COMBINING ENCLOSING KEYCAP} Set roles removed on mute (currently {0})".format(len(guild_settings.get("removed_on_mute_roles", set()))),
+                    "4\N{COMBINING ENCLOSING KEYCAP} Set moderator role (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("guild_moderator_role_id")) or MentionableNone("none")),
+                    "5\N{COMBINING ENCLOSING KEYCAP} Set custom role master (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("custom_role_id")) or MentionableNone("none")),
+                    "6\N{COMBINING ENCLOSING KEYCAP} Set custom role position (currently {0.mention})".format(ctx.guild.get_role(guild_settings.get("custom_role_position_id")) or MentionableNone("none")),
                 ])
 
             # Send embed
@@ -211,9 +212,10 @@ class BotSettings(utils.Cog):
             emoji_key_map = {
                 "1\N{COMBINING ENCLOSING KEYCAP}": ("verified_role_id", commands.RoleConverter, None),
                 "2\N{COMBINING ENCLOSING KEYCAP}": ("muted_role_id", commands.RoleConverter, None),
-                "3\N{COMBINING ENCLOSING KEYCAP}": ("guild_moderator_role_id", commands.RoleConverter, None),
-                "4\N{COMBINING ENCLOSING KEYCAP}": ("custom_role_id", commands.RoleConverter, "+Users with this role are able to make/manage their own custom role name and colour."),
-                "5\N{COMBINING ENCLOSING KEYCAP}": ("custom_role_position_id", commands.RoleConverter, "+Give a role that newly created custom roles will be created _under_."),
+                "3\N{COMBINING ENCLOSING KEYCAP}": ("setup removeroles",),
+                "4\N{COMBINING ENCLOSING KEYCAP}": ("guild_moderator_role_id", commands.RoleConverter, None),
+                "5\N{COMBINING ENCLOSING KEYCAP}": ("custom_role_id", commands.RoleConverter, "+Users with this role are able to make/manage their own custom role name and colour."),
+                "6\N{COMBINING ENCLOSING KEYCAP}": ("custom_role_position_id", commands.RoleConverter, "+Give a role that newly created custom roles will be created _under_."),
                 self.TICK_EMOJI: None,
             }
             def check(r, u):
@@ -240,13 +242,18 @@ class BotSettings(utils.Cog):
                 pass
 
             # And do some settin up
-            v = await self.set_data_in_guild_settings(ctx, key[0], key[1], prompt=key[2])
-            if v is None:
-                try:
-                    await message.delete()
-                except discord.HTTPException:
-                    pass
-                return
+            if len(key) > 1:
+                v = await self.set_data_in_guild_settings(ctx, key[0], key[1], prompt=key[2])
+                if v is None:
+                    try:
+                        await message.delete()
+                    except discord.HTTPException:
+                        pass
+                    return
+            else:
+                command = self.bot.get_command(key[0])
+                ctx.invoke_meta = True
+                await command.invoke(ctx)
 
     @setup.command(cls=utils.Command)
     @utils.checks.meta_command()
@@ -332,6 +339,88 @@ class BotSettings(utils.Cog):
                         ctx.guild.id, new_role.id, str(cooldown_time.duration)
                     )
                 self.bot.guild_settings[ctx.guild.id]['role_interaction_cooldowns'][new_role.id] = cooldown_time.duration
+
+    @setup.command(cls=utils.Command)
+    @utils.checks.meta_command()
+    async def removeroles(self, ctx:utils.Context):
+        """Talks the bot through a setup"""
+
+        message = None
+        while True:
+
+            # Construct embed
+            role_settings = self.bot.guild_settings[ctx.guild.id].get("removed_on_mute_roles", list())
+            with utils.Embed() as embed:
+                role_emoji = []
+                description = []
+                for i in role_settings:
+                    role = ctx.guild.get_role(i)
+                    if role is None:
+                        continue
+                    text = f"{len(role_emoji) + 1}\N{COMBINING ENCLOSING KEYCAP} {role.mention}"
+                    description.append(text)
+                    role_emoji.append((f"{len(role_emoji) + 1}\N{COMBINING ENCLOSING KEYCAP}", role))
+                embed.description = '\n'.join(description) or "No roles set up"
+
+            # Send embed
+            if message is None:
+                message = await ctx.send(embed=embed)
+            else:
+                await message.edit(embed=embed)
+                await message.clear_reactions()
+            if '@' in embed.description:
+                for i in range(1, len(role_settings) + 1):
+                    await message.add_reaction(f"{i}\N{COMBINING ENCLOSING KEYCAP}")
+            await message.add_reaction(self.PLUS_EMOJI)
+            await message.add_reaction(self.TICK_EMOJI)
+
+            # Wait for added reaction
+            emoji_key_map = {
+                self.PLUS_EMOJI: False,
+                self.TICK_EMOJI: None,
+            }
+            emoji_key_map.update({i:o for i, o in role_emoji})
+            def check(r, u):
+                return u.id == ctx.author.id and r.message.id == message.id and str(r.emoji) in emoji_key_map.keys()
+            try:
+                r, _ = await self.bot.wait_for("reaction_add", check=check, timeout=120)
+            except asyncio.TimeoutError:
+                return await ctx.send("Timed out setting up database.")
+
+            # See what our key is
+            emoji = str(r.emoji)
+            key = emoji_key_map[emoji]
+            if key is None:
+                try:
+                    await message.delete()
+                except discord.HTTPException:
+                    pass
+                return
+
+            # Try and remove their reaction
+            try:
+                await message.remove_reaction(r, ctx.author)
+            except discord.Forbidden:
+                pass
+
+            # They wanna delete a role
+            if isinstance(key, discord.Role):
+                self.bot.guild_settings[ctx.guild.id]['removed_on_mute_roles'].remove(role.id)
+                async with self.bot.database() as db:
+                    await db("DELETE FROM role_list WHERE guild_id=$1 AND role_id=$2 AND key='RemoveOnMute'", ctx.guild.id, role.id)
+                continue
+
+            # They wanna add a new role
+            if key is False:
+                new_role = await self.ask_for_information(ctx, "What role do you want to add?", "role", commands.RoleConverter)
+                if new_role is None:
+                    continue
+                async with self.bot.database() as db:
+                    await db(
+                        """INSERT INTO role_list (guild_id, role_id, key) VALUES ($1, $2, 'RemoveOnMute')""",
+                        ctx.guild.id, new_role.id
+                    )
+                self.bot.guild_settings[ctx.guild.id]['removed_on_mute_roles'].append(new_role.id)
 
     @setup.command(cls=utils.Command)
     @utils.checks.meta_command()
@@ -484,8 +573,7 @@ class BotSettings(utils.Cog):
 
         # Send message
         bot_message = await ctx.send(prompt)
-        def check(m):
-            return m.channel.id == ctx.channel.id and m.author.id == ctx.author.id
+        check = lambda m: m.channel.id == ctx.channel.id and m.author.id == ctx.author.id
         try:
             user_message = await self.bot.wait_for("message", timeout=120, check=check)
         except asyncio.TimeoutError:
